change to fuser instead of lsof; refactor disk files
